Update label for archived topics and add new topic
https://www.pivotaltracker.com/story/show/171338741

Below are two updates/additions to our Videos page on VSP.

We need to have a video channel added to our video page. Please add “Torque Factor” as a video schedule and header on this page:
https://www.vehicleservicepros.com/videos

The section should go between Tool Reviews and Kolman’s Korner

On the header for Kolman’s Korner” and “Toolbox Topics” please add “Archive” to the beginning, so both read as follows:
"Archive: Kolman’s Korner"

"Archive: Toolbox Topics"

(Both video series’ are no longer produced, but still good content for the site)

Acceptance Criteria: This ticket will be accepted when the Videos page on VSP reflect the two updates outlined in this ticket.