Update subscribe page description for LL
https://www.pivotaltracker.com/story/show/171362623

https://www.locksmithledger.com/subscribe

Please update it with the following:

Locksmith Ledger has products that deliver powerful content to you in a variety of forms including print, online, mobile and social media. To subscribe to a product or manage your current subscription, click on the icons below and put the power of Locksmith Ledger at your fingertips.

Acceptance Criteria: This ticket will be accepted when Locksmith Ledger's 'Subscribe' page has been updated with corrected copy.